# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Invoice Capture MVP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /source-documents:
    get:
      tags: [sourceDocuments]
      operationId: listSourceDocuments
      summary: List uploaded source documents with linked-invoice counts
      parameters:
        - name: processingStatus
          in: query
          schema:
            type: string
            enum: [PENDING, DETECTING, COMPLETED, EXCEPTION]
        - name: includeRemoved
          in: query
          schema:
            type: boolean
            default: false
        - name: search
          in: query
          schema:
            type: string
          description: Match by original file name
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Source document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDocumentListResponse"
    post:
      tags: [sourceDocuments]
      operationId: createSourceDocument
      summary: Create a source document from an uploaded file and start detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceDocumentInput"
      responses:
        "201":
          description: Source document created; detection started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDocumentWithInvoices"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /source-documents/{id}:
    get:
      tags: [sourceDocuments]
      operationId: getSourceDocument
      summary: Get a source document with its detected invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source document with invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDocumentWithInvoices"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [sourceDocuments]
      operationId: deleteSourceDocument
      summary: Permanently hard-delete a source document and all its invoices (test-data cleanup)
      description: >-
        Permanently removes a source document together with every child invoice
        and their audit/extraction records, and deletes the stored file. Blocked
        when any child invoice is posted. Requires explicit confirmation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HardDeleteInput"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Source document cannot be hard-deleted (e.g. posted child) or confirmation missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /source-documents/{id}/remove:
    post:
      tags: [sourceDocuments]
      operationId: removeSourceDocument
      summary: Remove a source document and void all its invoices (soft removal with reason)
      description: >-
        Marks a source document and all of its child invoices as removed/voided
        with a required reason. Removed records are excluded from active lists,
        queues, KPIs, CSV export and duplicate detection by default.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemovalInput"
      responses:
        "200":
          description: Source document after removal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDocumentWithInvoices"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Missing removal reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── Storage ─────────────────────────────────────────────────────────────────
...
  /source-documents/{id}/audit:
    get:
      tags: [sourceDocuments]
      operationId: getSourceDocumentAudit
      summary: Aggregated audit trail across a source document's invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"

  # ─── Phase 2: Imports (file-based reference data) ─────────────────────────────
...
components:
  schemas:
...
    Invoice:
      type: object
      required:
        - id
        - status
        - createdAt
        - updatedAt
        - fileObjectPath
        - originalFileName
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [PENDING_EXTRACTION, EXCEPTION, PENDING_APPROVAL, APPROVED, POSTED, VOIDED]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        poNumber:
          type: ["string", "null"]
        currency:
          type: string
          default: USD
        fileObjectPath:
          type: string
        originalFileName:
          type: string
        documentId:
          type: ["string", "null"]
        businessDocumentId:
          type: ["string", "null"]
          description: Business-facing display ID formatted as "VendorID - InvoiceNumber - Amount" (e.g. "V-00123 - INV-12345 - 1309.00"). Null until vendor is matched and invoice number/total are available.
        vendorRawName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        voucherId:
          type: ["string", "null"]
        exceptionReason:
          type: ["string", "null"]
        lowConfidenceFields:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        subtotal:
          type: ["number", "null"]
        freightAmount:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
          description: Discount/credit reduction extracted from the invoice (stored as a positive magnitude; subtracted in the tie-out formula).
        otherChargesAmount:
          type: ["number", "null"]
          description: Other charges, fees, or surcharges extracted from the invoice (added in the tie-out formula; negative values represent credits).
        tieOutExpectedTotal:
          type: ["number", "null"]
          description: Computed expected total = subtotal + tax + freight + other charges − discount.
        tieOutDifference:
          type: ["number", "null"]
          description: Extracted invoice total minus the expected total.
        tieOutStatus:
          type: ["string", "null"]
          description: Header tie-out outcome — one of PASS, WARNING, FAIL, SKIPPED.
        tieOutExplanation:
          type: ["string", "null"]
          description: Human-readable explanation of how the tie-out was computed and why it passed, warned, failed, or was skipped.
        paymentTerms:
          type: ["string", "null"]
        vendorMatchScore:
          type: ["number", "null"]
        validationStatus:
          type: ["string", "null"]
          description: Outcome of the authoritative validation engine (e.g. PASS, NEEDS_REVIEW, FAIL).
        reviewStatus:
          type: ["string", "null"]
          description: Review flag set when an invoice needs human review (e.g. NEEDS_REVIEW).
        overallReviewStatus:
          type: ["string", "null"]
          description: Aggregate human review state (e.g. PENDING, APPROVED).
        duplicateCheck:
          type: ["string", "null"]
          description: Result of the duplicate (vendor + invoice number) check.
        vendorCheck:
          type: ["string", "null"]
          description: Result of the vendor validation check (required/active/not on hold/matched).
        poCheck:
          type: ["string", "null"]
          description: Result of the purchase order capture check.
        amountCheck:
          type: ["string", "null"]
          description: Result of the amount validation check (total > 0, currency).
        totalTieOut:
          type: ["string", "null"]
          description: Result of the header tie-out check (subtotal + tax + freight = total).
        validationDetails:
          type: ["string", "null"]
          description: JSON detail of all validation checks, blocking issues, and warnings.
        extractionStatus:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          default: PENDING
        extractionError:
          type: ["string", "null"]
        extractionAttempts:
          type: integer
          default: 0
        extractionErrorDetail:
          type: ["string", "null"]
          description: Safe JSON troubleshooting detail for a failed extraction (no sensitive data).
        fieldConfidence:
          type: ["string", "null"]
          description: JSON map of per-field confidence (0-100) keyed by field name.
        extractionNotes:
          type: ["string", "null"]
        lastExtractedAt:
          type: ["string", "null"]
          format: date-time
        sourceDocumentId:
          type: ["integer", "null"]
          description: ID of the source document this invoice was detected/split from.
        invoiceSequence:
          type: ["integer", "null"]
          description: 1-based position of this invoice within its source document.
        pageStart:
          type: ["integer", "null"]
          description: 1-based first page of this invoice within the source document.
        pageEnd:
          type: ["integer", "null"]
          description: 1-based last page (inclusive) of this invoice within the source document.
        role:
          type: string
          enum: [AP_PROCESSOR, AP_APPROVER]
          default: AP_PROCESSOR
        removedAt:
          type: ["string", "null"]
          format: date-time
          description: Timestamp when the invoice was voided/removed. Null when active.
        removedBy:
          type: ["string", "null"]
          description: Actor who voided/removed the invoice.
        removalReason:
          type: ["string", "null"]
          description: Required reason captured when the invoice was voided/removed.
        removalNote:
          type: ["string", "null"]
          description: Optional free-text note captured at removal time.
        exportStatus:
          type: ["string", "null"]
          description: File-based export readiness — one of NOT_READY, READY, EXPORTED, FAILED, BLOCKED. This is NOT an ERP status.
        exportBatchId:
          type: ["string", "null"]
          description: Business ID of the export batch this invoice was last included in.
        exportedAt:
          type: ["string", "null"]
          format: date-time
        exportBlockedReason:
          type: ["string", "null"]
        exportRetryCount:
          type: integer
          default: 0
        exportFileName:
          type: ["string", "null"]
        exportFormat:
          type: ["string", "null"]
        exceptionOwner:
          type: ["string", "null"]
          description: AP user assigned to work this exception.
        exceptionReviewedAt:
          type: ["string", "null"]
          format: date-time
        exceptionReviewedBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SourceDocument:
      type: object
      required:
        - id
        - originalFileName
        - fileObjectPath
        - sourceChannel
        - processingStatus
        - uploadedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        originalFileName:
          type: string
        fileObjectPath:
          type: string
        fileHash:
          type: ["string", "null"]
        sourceChannel:
          type: string
        uploadedBy:
          type: ["string", "null"]
        uploadedAt:
          type: string
          format: date-time
        pageCount:
          type: ["integer", "null"]
        detectedInvoiceCount:
          type: ["integer", "null"]
        processingStatus:
          type: string
          enum: [PENDING, DETECTING, COMPLETED, EXCEPTION]
        processingError:
          type: ["string", "null"]
        removedAt:
          type: ["string", "null"]
          format: date-time
          description: Timestamp when the source document was removed. Null when active.
        removedBy:
          type: ["string", "null"]
          description: Actor who removed the source document.
        removalReason:
          type: ["string", "null"]
          description: Required reason captured when the source document was removed.
        removalNote:
          type: ["string", "null"]
          description: Optional free-text note captured at removal time.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SourceDocumentInput:
      type: object
      required:
        - fileObjectPath
        - originalFileName
      properties:
        fileObjectPath:
          type: string
          minLength: 1
        originalFileName:
          type: string
          minLength: 1
        contentType:
          type: ["string", "null"]
        fileHash:
          type: ["string", "null"]
    SourceDocumentWithInvoices:
      type: object
      required:
        - source
        - invoices
        - invoiceCount
        - extractedCount
        - exceptionCount
        - pendingCount
        - removedCount
      properties:
        source:
          $ref: "#/components/schemas/SourceDocument"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        invoiceCount:
          type: integer
          description: Count of active (non-removed) invoices.
        extractedCount:
          type: integer
        exceptionCount:
          type: integer
        pendingCount:
          type: integer
        removedCount:
          type: integer
          description: Count of voided/removed invoices belonging to this source document.
    RemovalInput:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          description: Required reason for voiding/removing the record.
        note:
          type: ["string", "null"]
          description: Optional free-text note.
        actor:
          type: ["string", "null"]
          description: Actor performing the removal (permissions deferred).
    HardDeleteInput:
      type: object
      required:
        - confirm
      properties:
        confirm:
          type: boolean
          description: Must be true to perform a permanent hard delete.
        actor:
          type: ["string", "null"]
          description: Actor performing the deletion (permissions deferred).
    DeleteResult:
      type: object
      required:
        - deleted
        - deletedInvoiceIds
        - fileDeleted
      properties:
        deleted:
          type: boolean
        deletedInvoiceIds:
          type: array
          items:
            type: integer
        deletedSourceDocumentId:
          type: ["integer", "null"]
        fileDeleted:
          type: boolean
          description: Whether the underlying stored file was also deleted.

    # ─── Vendor ─────────────────────────────────────────────────────────────────
...
    AuditLogEntry:
      type: object
      required:
        - id
        - invoiceId
        - action
        - createdAt
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        action:
          type: string
        fieldName:
          type: ["string", "null"]
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        editorRole:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── Storage ────────────────────────────────────────────────────────────────
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Phase 2: Dashboard & Analytics ──────────────────────────────────────
...
    SourceDocumentSummary:
      type: object
      required:
        - source
        - invoiceCount
        - extractedCount
        - exceptionCount
        - removedCount
      properties:
        source:
          $ref: "#/components/schemas/SourceDocument"
        invoiceCount:
          type: integer
        extractedCount:
          type: integer
        exceptionCount:
          type: integer
        removedCount:
          type: integer
    SourceDocumentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SourceDocumentSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── Phase 2: Imports ────────────────────────────────────────────────────
...